requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
google-cloud-bigquery>=3.10.0
python-dotenv>=1.0.0
typer[all]>=0.9.0
tqdm>=4.65.0
pyyaml>=6.0
black>=23.0.0
ruff>=0.0.285
pytest>=7.0
